requirements-airflow.txt
sqlalchemy>=1.4.28,<2.0
pymysql==1.1.0
mysql-connector-python==8.3.0
pandas==2.0.3
scikit-learn==1.3.2  # ← VERSIÓN FIJA
numpy==1.24.3
joblib==1.3.2
